UsuarioService

+++ src/main/java/br/com/codenation/aceleradev/rec/loja/entidades/Produto.java
@@ -1,10 +1,12 @@
 package br.com.codenation.aceleradev.rec.loja.entidades;
 
+import java.math.BigDecimal;
+
 public class Produto {
 
   private long id;
   private String descricao;
-  private double preco;
+  private BigDecimal preco;
   private int qtdEstoque;
   private int categoria;
 
@@ -24,11 +26,11 @@ public class Produto {
     this.descricao = descricao;
   }
 
-  public double getPreco() {
+  public BigDecimal getPreco() {
     return preco;
   }
 
-  public void setPreco(double preco) {
+  public void setPreco(BigDecimal preco) {
     this.preco = preco;
   }
